Suppress unused fields warnings when `alloc` feature is disabled
They are used only from URI resolution algorithm, so they are only
needed when `alloc` feature is enabled or tests are being compiled.

Defining the type when only needed is an option, but I didn't choose
that way since it would make things unnecessarily complicated.
Impact on the compilation time of `RiReferenceComponents` would be
quite little (I haven't measured though) and it does not require any
additional dependencies, so I feel it is not worth adding
`#[cfg(any(test, feature = "alloc")]` to the many place and maintaining
them.