Rewrite the relative IRI resolver
The new resolver is available for environment without `alloc`.

This change removes `is_strict` parameter from resolver, since the
strictness is related to the interpretation of the IRI string and it
is the parser matter. The parser provided by this crate is "strict",
so resolver should use an algorithm for the strict parser.