Rewrite parsers for more limited use cases
Previously, parsers are implemented using `nom` crate.
Such parsers are powerful so that "eat" valid input as much as possible
from any input that may have invalid garbage suffix.
However, use case of parsers (and validators) in this crate does not
need such power. The parsers are required to judge whether the complete
string is valid or not. Noticing "prefix is valid" is unnecessary.

Thus I decided to rewrite the parser for more limited situations.
Now the `parser` is split to modules:

* `char`: common functions for characters.
* `str`: common functions for strings.
* `trusted`: parsers for trusted input.
* `validate`: parsers for non-trusted input.

`char` module has functions to check characters' properties. The
functions are split into `ascii` version and `nonascii` version, since
IRIs consists of ASCII characters in many cases, even when it is an IRI
and contains some non-ASCII characters.
Most of predicates for ASCII characters are implemented as table
lookups, so they are expected to be quite fast.

`str` module has common generic functions to validate, search, and split
strings. They would be easily made faster by `memchr` crate, and I'll
introduce such improvements later.

Parsers in `trusted` module make an assumption that the input is already
validated (and actually valid), and the parser is used in correct way.
They fully utilize knowledges of IRI syntaxes and omit checks as
possible, so they are fast but they can be (and should be) used
internally.

Parsers in `validate` module can be used to validate any string input.
They strictly check that the input is valid, so they are slow but they
can be used for strings from non-trusted data source.

I tried some simple benchmarking, and this change makes IRI validation
and resolution roughly 3x to 4x faster.