Stop forcing `serde-{std,alloc}` when both serde and std/alloc is enabled
Previous implementation of this crate emits compiler error when both
of `serde` and `std`/`alloc` is enabled but combined
`serde-{std,alloc}` is disabled. However, `serde` and `std`/`alloc`
features could be enabled indirectly from different indirect
dependencies, so this is not desirable.

`serde` does not conflict with `std`/`alloc`, so it should not be
compile error when both are enabled.