Revert resolution base type change
RFC 3986 section 5.1. says:

> A base URI must conform to the <absolute-URI> syntax rule (Section
> 4.3). If the base URI is obtained from a URI reference, then that
> reference must be converted to absolute form and stripped of any
> fragment component prior to its use as a base URI.

I previously had thought that it would be convenient if `IRI` can be
passed and the resolver automatically strips the fragment component.
However I now changed my mind. The fact that the fragment component
of the base is stripped **even when the reference is empty** is not
so obvious for everyone, so it should be made explicit by letting
users write `.to_absolute()`.