Remove `NormalizationTask::write_to_iri_string`
It is the combination of reuse of the internal `String` and omission
of a redundant validation. This API is not so beautiful and would be
hard to use since it takes ownership from the given IRI string.

Alternative method could be introduced, for example, receiving `String`
or `Vec` directly and returning `Self::OutputOwned`. For now, however,
remove this method to make task-related refactoring easier.