Make `NormalizationTask` string type dependent
Normalization result should be more specific dependent to the input
IRI string type. For example, allocating variation of the normalization
should return `RiString<IriSpec>` for `&RiStr<IriSpec>`, but should
return `RiAbsoluteString<UriSpec>` for `&RiAbsoluteStr<UriSpec>`.
To implement this, keeping `S: Spec` as a type parameter of
`NormalizationTask` is not enough.