Stop attempting to follow WHATWG spec and add more normalization
Previous implementation says "normalization algorithm defined in WHATWG
URL Standard" but it does not actually follow the standard. When the
authority component is absent and the path is relative, the path should
be treated as "opaque" and dot segments removal should be avoided.

> 9. Otherwise, set _url_'s path to the empty string and set state to opaque path state.
>
> --- <https://url.spec.whatwg.org/#scheme-state> (as of 2022-09-13) or
> <https://web.archive.org/web/20220913000748/https://url.spec.whatwg.org/#scheme-state>
> (archive of the spec as of 2022-09-13).

I'm not sure I can correctly follow the WHATWG URL spec, and I'm not
going to strictly follow the spec (since the primary target of the crate
is RFC 3986, not the living standard by WHATWG), so I decided to stop
describing they are implemented according to WHATWG URL Standard.